Adjust Log Group Subscription Filter
Using the JSON notation of `{}` excludes all non-structured log lines
which occur when the tasks are starting up and when running migrations.
These log lines are important for debugging. Using this basic pattern
match works. It could be scoped tighter ideally with:
'-"\"action\": \"/ping\""'
but I've been unable to apply that pattern via Terraform and
its provided character escaping.