Reduce minimum capacity to 6 in production
Monitoring has shown us that 9 instances in production is quite a lot too many
for idle times. However, load tests starting with 3 instances at idle shows us
that 3 instances is too few to handle the loads with adequate response times
whilst the autoscaler kicks in. Thus, we settle for a minimum of 6 instances
when idle.