Enable AWS WAF to block IPs past a requests threshold
Previously we had AWS WAF in "count" mode, where it would tell us if it _would_
have blocked something. We are now putting it into block mode, where it _will_
block things.

In addition, we add an exception for the egress IPs of our NAT gateways, so
that we don't accidentally rate limit ourselves when the applications talk to
one another through the front door.