Persist gems between service restarts
In a previous commit we started installing development gems into the
forms-admin container as part of the start command. These gems aren't
present in the base image, so this means everytime the service is
started we have to start from scratch, which takes some time.

This commit uses a volume to persist the directory where gems are
installed, so we can save gems we've already installed from having to be
installed again after restarting the service. This reduces startup time
and network traffic.

Although this is only really an issue for the forms-admin service
currently, for consistency and because there's no downside we make the
same change to all forms app services.