Allow email domains for potential partners access to GOV.UK Forms
We have some partners we're working with, but who don't have access to
the platform because their organisation doesn't have an email domain
that ends with `.gov.uk`. We want to allow these people in, as they'll
be needing to use the platform soon.

This commit adds the relevant email domains to the list of domains
allowed by our Auth0 pre-registration action.

This commit only allows access to production system; we prefer not to
give users outside of the GOV.UK Forms team access to the staging and
development systems [[1]].

[1]: https://github.com/alphagov/forms-team/wiki/How-to-access-GOV.UK-Forms#our-environments