Use db:prepare instead of SQL, setup, seed, migrate
The `db:prepare` task takes care of all database setup stuff in one
command. We can simplify the init.sql script for setting up the
PostgreSQL database, as it will create the database if it doesn't
already exist. We don't need to call the `db:seed` or `db:migrate` tasks
as it will also take care of this.

Note that we remove from `init.sql` the lines about granting privileges
to the database, but they are unneeded as the user we give the user we
create for each app the ability to create databases, and when a user
creates a database they will automatically be the owner and have all
privileges already [[1]].

[1]: https://www.postgresql.org/docs/current/manage-ag-createdb.html