Use default database name for apps
The configuration for the apps forms-admin and forms-api already include
a default database name in `config/database.yml` [[1], [[2]]. Removing
the database path from the DATABASE_URL environment variable tells Rails
to use this default. The advantage of using this default is a) less
configuration needed in the Compose file b) the same database can be
used for apps run in Compose or locally.

[1]: https://github.com/alphagov/forms-admin/blob/dbf27befd69f8ceef2e23a902afb24f0a9ac5db3/config/database.yml#L26
[2]: https://github.com/alphagov/forms-api/blob/76add6c6fb92a68ab2aba38690123760af676570/config/database.yml#L26