Separate pen tester CIDR ranges from VPN CIDR ranges
When we set up pen tester access, we added their CIDR ranges to the list of
allowed ranges next to those of our VPN. This has the side effect of adding
those ranges as allowable to all of our users, meaning we couldn't apply new
changes in productino without unnecessarily widening access.

This commit separates the two concerns and allows them to be set per
environment.