Use the same PostgreSQL user for all apps
Using the same SQL user for all apps makes it easier to use the same
database tables for local development and Docker Compose. This commit
changes all the Rails apps to use the default `postgres` user that comes
with the postgres Docker image.

For production systems it is good to give apps limited privileges to the
database, but for our development stack that isn't so much of an
issue.