Set the GIT_SHA env var when building the end-to-end tests container
The builds of this container have been failing because the resulting tag looks
like "-1234567", which is an invalid tag. The Git commit SHA can only be set as
an environment variable in the pipeline, and not as a Terraform variable to the
module, becuse the commit SHA isn't known until runtime.