Add validation to ensure correct case for allowed email domains
Auth0 appears to do case-folding to email addresses; when in an action
the email address of the user trying to sign in is always all lowercase,
regardless of how they typed it. This means our allowed email domains
also need to be all lower case, so that the comparison in the JavaScript
can compare correctly.

This commit adds a validation to enforce that the strings in the list of
allowed email domains are all lowercase. We could do the casefolding in
the action, but this does the computation up-front so is faster when it
comes to users actually signing in.