Increase defult Auth0 OTP expiry time to 900s
A 3 minute expiry time is low enough that it's too easy for users to let their
codes expire for a whole host of reasons. 15 minutes (900s) is a nice balance
between not having super long-lived OTPs and giving users plenty of time to
make use of them.

We change only the default here so that any consumers of the module which have
set their own timeout can keep it.