Update content for Auth0 signup and login pages
We want to inform new users that they need a government email address
to create an account. This commit updates content in the description
for the create account page, and consolidates some titles and
descriptions into single lines. To do this, some page descriptions are
set to a single space, as empty strings results in Auth0
defaults being used, and zero-width spaces result in line breaks being
inserted.