Use node_modules folder from container, not from Docker host
As a general rule we want to keep any Node.js modules installed within a
Docker container separate from modules installed with the Docker host.
Because the `node_modules` folder for an app is always in the app
project directory, and we've used a bind mount to import the app
directory from the host for each service, we need to add another volume
on top to keep our node_modules folders separate.