Install Node.js development dependencies in app containers
We want to be able to use development dependencies such as Vite.js in
our local development stack, however the `Dockerfile` for each app
removes the development dependencies after the frontend assets have been
built.

This commit adds an invocation of `npm install` to the start command for
each app service, and ensures that NODE_ENV is set to `development`, so
that development dependencies are installed.