Allow the speed at which new continers are started on ECS to be customized
When a AWS ECS is deploying a new task definition, it will start a number of
new instances, and wait for them to be healthy before draining a corresponding
number of old instances. The maximum number it will deploy at once is given as
a percentage of the currently running number.

Historically we've set this to 200% so that ECS can deploy the total number
required all at once. We've done this to speed up deployments overall.

We'd now like to be able to let individual services control the rate at which
thier instances are rolled out. The default is left at 200% so that consumers
that don't want to set the value will continue to see the same behaviour.