Deploy forms-api and forms-runner changes 34% at a time
When we tested performing a rolling deplyoment of forms-api and forms-runner
during a high load scenario, we found that deploying one would knock out the
other. We believe this is because the default 100% extra instances being
deployed for a period of time was causing an outsized load on the other
component, which in turn caused it to fail.

We believe we can address this problem by not deploying 100% of the new
instances at once.

We're spread across 3 availability zones, and we enforce that instance counts
must be multiples of 3, so deploying 1/3 at a time is a good bet (traded off
against increased deployment times).

34% is not 1/3, I know, but 1/3=0.333 and ECS only allows whole numbers. So 34%
it is.