Define a set of inputs for all Terraform deployments
Introduces a set of variables which can, in future commits, be integrated into
every Terraform deployment. Also included is a per-environment-type set of
variable definitions.

These files are not integrated into anything at the moment on purpose. Shortly
after this, we will be flattening our Terraform deployments in to one set which
we will parameterise using the files in this commit.