Extract account-level configuration to its own deployment
Previously we configured account level things such as contact details, Route53
hosted zones, and IAM roles at the same level as the deployment of the GOV.UK
Forms product. In the future we want to be able to deploy multiple environments
into the same accounts, without duplicating the account level stuff (which
potentially results in conflicts). The answer to this problem is to extract
everything that operates at the account level to its own deployment, which we
expect will never be automated, and only ever be run by a human.

An added benefit is the simplification of DNS delegation.

More will be written elsewhere about how this changes affects us as operators,
and how it changes how we deploy environments.