Upgrade to Terraform 1.7-rc1
We will shortly be merging a larger change to our Terraform structure, in which
we require features from 1.7. We want to upgrade everything else prior to
merging the structural change, so that we have matching versions in the event
we need to make any changes to Terraform before we merge.