Make action of the Auth0 root conditional
In the user-research environment we use HTTP basic auth for authentiction, not
auth0. We don't (currently) have the quota spac available in Auth0 to begin
using it either.

In `main.tf` we have to use a "count" argument on the module invocation to
disable the module on a condition. This means the Terraform resource name will
change from "module.auth0" to "module.auth0[0]". To prevent Terraform from
destroying and recreating the resource, we use a "moved" block to tell
Terraform about the intentional change of name.

In `site.tf` we have to conditionally look up the Auth0 credentials from AWS
SSM. This is because the credentials won't exist in the user-research
environment, and attempting to find them will cause an error in Terraform. The
workaround is inelegant, but it's concise enough to be maintainable, hopefully.

Once this code has been run once in each environment, the "moved" block can be
removed.