Remove extra checks for Checkov
With the new terraform structure our checks are obsolete because Checkov is not
able to look up the values from the `tfvars` files. Even if it did, we would
need it to do the same for each of the environments.

We will remove these extra checks for now and work on a way to validate these
values later :(