Ignore RDS version changes, so that we don't downgrade accidentally
We specified version 11.18 when we created the database clusters but since then
AWS have provided automatic minor version upgrdes.

We don't wish for Terraform to attempt to downgrade the engine version, or to
have to update our config every time there's a new minor version.

Instead, we ignore any changes to the engine version, and allow AWS to be the
arbiter of the exact version.

When we want to perform major version upgrdes, we can remove this lifecycle
configuration, and replace it once the upgrade is complete.