Add Node.js to update_app_versions script
Alpine Linux tends to have only one version of Node.js in its main repos
at a time. It would be helpful if this version was the same as we use
for local development and continuous integration.

This commit adds functionality to the `update_app_versions` script to
get the Node.js version from the base image and add it to the `.nvmrc`
file.