Extract Terraform invocation knowledge into a script
Our Terraform root modules are beginning to require specific arguments in order
to be run correctly. We originally encoded this knowledge directly in the
Makefile targets, but now we want to add other deployments which will need
other flags. To address that, and keep the Makefile from growing overly
complex, we've extracted that knowledge into a separate script file which is
invoked by Make.