Extract CodeBuild IAM configuration from the code-build-build module
The deployer-access module lists explicitly which roles from which accounts can
assume the deployer role, therefore we need to be able to succinctly describe
the roles being used by pipelines which are deploying into an account.
Previously, the code-build-build module created one role per invocation, which
meant an explosion in roles which were identical but for name. To avoid that,
we've extracted the IAM configuration and it is now the job of the caller to
provide a role that has sufficient permission for its needs.

Similarly, we've extracted the desired name of the log group to the caller,
because the role must be able to access that log group and the only way for it
to reliably know the name of the log group is to be the one specifying it.

Do not deploy the deploy-access root in the pipelines:

We have moved this root to the `account` deployment. The deployer role has a lot
of power and we want to make sure any changes to it have to be applied by a
human, not a pipeline