Move deployer-access root to account deployment
We have a deployer-access role per environment but they are all the same (except
for their names). We want to create it as a single role and set the permissions
at the account level. As a first step we have moved the deployer-access root to
be part of the account deployments. Given that the deployer-access role has a
lot of power, we want to make sure it is only human-deployable.

In future PRs we will make sure only one role exists for all environments,
rather than one per environment.