Merge policy documents
AWS limits the number of policies per role (to 10) so we have to merge these
policy documents into one so that we can we can create a single policy to attach
to the role.

AWS also limits the size of a policy (PolicySize: 6144) so we need multiple of
them BUT FEWER THAN 10. I am fine.

It might make sense to structure these differently, but for now this will work