Add AUTH_KEY environment variable to forms-api
We want to rename the config setting forms_api.authentication_key to
forms_api.auth_key in the forms-api app, to match the other apps.

To do this without breakage we are going to use the expand-contract
pattern.

This commit adds a new environment variable with the new name. Once all
instances of the forms-api app are using the new environment variable,
we can remove the old one.