Upgrade Aurora PostgreSQL to version 13.12
Version 13.2 is the newest version we can upgrade to, from the current version,
at the time of writing.  This was found by running

aws rds describe-db-engine-versions \
  --engine aurora-postgresql \
  --engine-version "11.serverless_21" \
  --query 'DBEngineVersions[*].ValidUpgradeTarget[*].{EngineVersion:EngineVersion}' \
  --output text

The Terraform lifecycle rule ignoring changes to the version has been commented
out so that the version can be changed. Once it has been deployed, we will
restore the rule.

Also sets
* 'allow_major_version_upgrade = true' to allow the major version to
be upgraded

'allow_major_version_upgrade' will be reverted afterwards.