Enable IAM Access for kms key
To manage a KMS key, both the key policy and the policies attached to the
principal (user/role/group) must have a statement allowing management of the
key. Having this very permissive policy means one of the two sides will always
allow it. This reduces the risk of the key becoming unmanageable [1]

[1] https://docs.aws.amazon.com/kms/latest/developerguide/key-policy-default.html#key-policy-default-allow-root-enable-iam