Make it possible to apply RDS changes immediately
By default we want all changes to RDS to be applied in the next maintenance
window, but in some cases we'd like to be able to override that behaviour and
have them take effect immediately (e.g. while we test upgrade dev).

This can now be done by setting the 'TF_VAR_apply_immediately' environment
variable to 'true' (default is false), and there is a message reminding operators of this when
they apply/plan the 'forms/rds' root.