Ensure all Forms environments can read from GitHub
We a CodeStar connection to allow us to read from GitHub. Unfortuantely, the
creation of this connection can't be automated because it requires an OAuth 2
authorization code flow, which needs a human in the mix.

To make up for that, we create it by hand once in each account and record its
ARN in tfvars.