Introduce 'deploy-forms-product-page-container'
This commit is quite large so we'll break down what's going
on.

1. Introduces a nwe pipeline called 'deploy-forms-product-page-container'
2. Configures the pipeline to be triggered by events from AWS EventBridge

As of this commit, this is only configured to target a dev env, but future
commits will make this applicable in every environment, as well as attempt to
make the pattern more reproducible.

-- 1
We have a new pipeline: 'deploy-forms-product-page-container'. The goal of this
pipeline is to deploy new versions of the GOV.UK Forms product pages container
to AWS ECS, in response to the existence of new tags matching a pattern in the
AWS ECR repository in which it lives.

There are two challenges faced by this pipeline.

First, it needs to be triggered by the existence of a new tag in the ECR
repository, but the CodePipeline ECR action cannot work with repositories in
other accounts (such as we have). To work around this, we take the image URI
input from a variable. The value of the variable is provided by a human when
triggering it from the console, and by EventBridge when being triggered
automatically (see 2).

Second is the resultant lack of a traditional source, such as a git repository.
This is a problem because AWS CodePipeline strictly requires at least one
source action in every pipeline, and all CodeBuild actions require an input
artifact (containing their buildspec) - even though in this instance our
CodeBuild actions doesn't need any input beyond environment variables. We could
have solved this by pulling in the forms-deploy repository, but that would have
introduced a new problem: what version should we get? Instead, we have opted to
introduce as minimal a source as possible. We use the CodePipeline artifact
bucket (S3) to store a zip file containing the buildspec for a later CodeBuild
action, and tell CodePipeline to pull its source code from that object. This
zip file gets updated every time we re-deploy the pipelines, which is
sufficient to keep it up to date and in sync with the pipelines.

-- 2
Since the AWS CodePipeline ECR action can't work with repositories in another
account, we have to work around it. Outside of this commit, we have configured
AWS EventBridge to forward AWS ECR events from the account containing the
registry to this account. In this commit, we set up a new rule to detect pushes
of new tags matching a pattern in the desired repository. Once detected, AWS
EventBridge invokes the new pipeline using its input transformation feature[1].
We configure the IAM role used by EventBridge to allow for this.

[1]: https://docs.aws.amazon.com/eventbridge/latest/userguide/eb-transform-target-input.html