Use deployer role as default role when running smoke tests
Rather than creating a role specifically to run the smoketests, we have decided to use the deployer role to do this by default. We still have a specific role to do this in the terraform, but this will only be created if the deployer role does not exist. In order to use the deployer role for this, I have added some permissions relating to ecs and ssm