Allow deployer roles full access to ECR in deployer account
The deployer role will eventually need to be able to authenticate, pull, and
push images across a number of repositories in the 'deploy' account. This will
lead to lots of IAM policies on an already stretched role if managed
separately. To avoid that, we grant the role access to do 'ecr:*' in the
'deploy' account, and manage the access in a more fine-grained way using
resource-based policies on the repositories themselves.