Promote forms product page image at the end of the pipeline
Image promotion is achieved by pulling the image which was just successfully
deployed and tested, and re-tagging it, and pushing it back to the repostiory.
The result is one image with two tags, the presence of which will trigger
further pipeline runs.

We use a sed [1] expression, with the current tag as input, to generate the new
tag. We do this so that the pipeline does not need to have a copy of the git
repository in order to maintain the commit hash as part of the tag. Instead, we
use sed to effectively swap the prefix.

[1] https://www.gnu.org/software/sed/manual/html_node/sed-Programs.html#sed-Programs