Introduce image builder pipelines
We have pipeline definitions which can, on the existence of an image tag,
deploy the image, test it, and promote it all the way to production. This
commit introduces a new pipeline which builds a new container and tags it to
begin the former process when a new commit is added to the 'main' branch of a
repository.