Add digitalaccessibilitycentre.org to staging
To give us a quote for an accessibility audit of our
application the testers need to access our system.

This commit adds the domain digitalaccessibilitycentre.org to the
allowed domain names in the staging environment to that they can create
a trial user.

The staging environment was chosen as:
- It's less risky than production but more stable than dev
- We did the same for the pentesters who needed access

This will allow all users with email accounts ending in
digitalaccessibilitycentre.org to create accounts. I think that's the
right choice as it's a fairly small group and doesn't look like it hands
out email addresses to anyone else.

After the audit is costed, we can decide if we need to keep this domain
on the list so the audit can be completed or if it can be removed.

```bash
gds aws forms-staging-readonly -- make staging forms/auth0 plan
```

This is only a partial output because stagign is currently deployed with
auth0 settings from another branch.

```
~ resource "auth0_action" "restrict_users_to_allowed_domains" {
    ~ code       = <<-EOT
          /**
          * Handler that will be called during the execution of a PreUserRegistration flow.
          *
          * @param {Event} event - Details about the context and user that is attempting to register.
          * @param {PreUserRegistrationAPI} api - Interface whose methods can be used to change the behavior of the signup.
          */
        - let domains = [".gov.uk"]
        + let domains = [".gov.uk","@digitalaccessibilitycentre.org"]
          exports.onExecutePreUserRegistration = async (event, api) => {
            if (event.user.email && !domains.some((domain) => event.user.email.endsWith(domain))) {
              api.access.deny("unauthorised_email_domain", "Use your government email address.");
            }
          };
      EOT
      id         = "b9986e8e-eede-4d78-9a5c-c04f2c157e03"
      name       = "restrict-users-to-allowed-domains"
      # (3 unchanged attributes hidden)

      # (1 unchanged block hidden)
  }
```