Ensure the externally-managed Terraform Auth0 connection is maintained
We've gone from a model of appending to a list to one of providing the list
idempotently. This means we need to always know all of the values that the list
needs to contain, which in this case means knowing a value that is managed
outside of Terraform.

We've added a new input to the Auth0 module to accept the extra values.