Use Makefile targets when applying Terraform in pipelines
Our Makefile targets make our intent very clear, and the script they call to
invoke terraform encodes a lot of important information about how to invoke
different pieces of Terraform. We don't want to spread that information around,
so we should invoke that script in places where we need it.