Enable the ecs-service module to maintain the current image
Historically, the ecs-service module has required an image URI as input so that
it can inform AWS ECS on Fargate which container image to use.

However, we now have need for it to be able to deploy changse to ECS
configuration without changing the running container image, and without being
given that value as input.

This comes because we're moving to deploy ECS changes in two ways:
1. Changes to the ECS configuration (allocated memory, subnet placement etc)
2. Changes to the running images

In the former, we don't want to have to tell Terraform what the currently
running image is (it can divine that for itself). In the latter, we don't have
(and don't want) access to the `forms-deploy` repository to run this Terraform.

We have maintained the ability to override the running container to support
first-time deplyoments where there isn't a running container to look up.