Replace EventBridge->CodePipeline with EventBridge->Lambda->CodePipeline
AWS EventBridge documentation says that a target with an input transformer will
receive the results of the input transformer as its input. However, we did not
observe this to be true when the target ARN was a pipeline. We raised a support
ticket with AWS, who confirmed that this is a "known issue" and the two
services don't work together like that.

Meanwhile, we still need this to be able to invoke an AWS CodePipeline pipeline
with specific input in reaction to an event. To work around this issue, we add
a very small, single-purpose AWS Lambda function which accepts valid
`codepipeline:StartPipelineExecution` inputs and proxies it to CodePipeline.

It is written in Ruby and has a few dependencies as reasonably possible, so
that it incurs as little maintenance overhead as possible.

Once AWS fixes this integration, we can rip this extra step out.