Conditionally disable e2e tests in deploy-product-page-container pipeline
It isn't possible to run the end-to-end tests in the user-research environment
because it doesn't have Auth0 configured.

CodePipeline doesn't support conditional execution, or a lightweight way to
wrap something in little bit of Bash, and making the end-to-end tests module
responsible for skipping itself with a flag feels wrong -- so conditionally
removing the stage from the pipeline in Terraform seems the best solution.