Apply 'pipelines' and 'forms-product-pages' Terraform automaticaly
We have a pipeline which applies the different Terraform roots. It did not
previously apply its own Terraform, but it should.

It also did not apply the forms-product-page Terraform because of the historic
need to know which container image to use. This is a solved problem in the
forms-product-page root now, so we can automatically apply it.