Grant 'deployer-$ENV' role full read-only access
The deployer role works with a lot of different services and APIs, so it
naturally needs to be able to read a lot of things. Reading in itself is a low
risk action (especially when adding in resource-based policies as an additional
layer of protection), so we can reduce the footprint of our policies by
granting the role full read-only access, and focussing our efforts on
describing the creates, updates, and deletes that the role needs be allowed to
perform.