Use a Forms-specific event to promote forms-deploy commits through envs
When CodePipeline emits an event to say the pipeline ran to completion it
doesn't publish the commit ID that it ran with. This is a problem for us
because we want the next stage(s) to use the same commit ID, and not whatever
is at the tip of main in that moment.

To work around that, we publish a custom event at the end of the pipeline that
contains the information we need, and each stage reacts in the same ways as
before.