FOR SAFETY: plan instead of apply Terraform in automatic pipelines
When we merge this into `main` we expect that the staging environment will pick
up the changes and begin rolling them out. To be safe, we'd like it to plan the
changes instead of applying them on the first roll through, just in case.

Once we're happy with it, we can revert this commit.