Fix "server already running" error
Occasionally app services in our Compose stack will fail to start with
the error:

```
A server is already running. Check /app/tmp/pids/server.pid.
```

We tried to fix this issue in commit ef9dfc1d, but that change hasn't
completely eliminated occurrences of the error.

This commit copies the solution from govuk-docker [[1]], by adding the
`--restart` flag to `rails server`. As noted on the govuk-docker commit,
this flag is primarily inteded for use by the `rake restart` task, but
as it also cleans up the pidfile it does fix the issue we're seeing.

[1]: https://github.com/alphagov/govuk-docker/commit/45ef6f845eb82e2657861555fc99cf9774b1e20d