Reinstate sending CodePipeline events to the deploy account event bus
The deploy account has an AWS ChatBot configuration which sends CodePipeline
events to Slack. Whilst it might be nice to have this configured on a
per-environment basis, AWS ChatBot does not have an API so we can only clickops
it, and it'll be better to use a central configuration for as long as that's
true.