Send notifications to Slack when pipeline statuses change
We have notifications via CodeStar Connection Notifications for pipelines in
the deploy account (which will eventually be retired), but we also need
notifcations for pipelines in other accounts.

We use AWS ChatBot for sending messages to Slack, but that doesn't have an API
and we've had to manually configure it in the AWS Console. For that reason, we
only want to have one configuration of it globally, so we need to send all of
our notifications through it.

AWS CodeStar Connection Notifications sends messages to Slack by sending a
message to an AWS SNS topic with a known payload format. We can do the same
thing to achieve the same result using AWS EventBridge.