add policy to manage deployments to support role in all environments
Previously, our pipelines were only in the deploy account, and so we did not need to add any permissions to manage the pipelines to the environment accounts. Now that our environment accounts will have pipelines, we need to update the support role to include the correct permissions so that we can manage deployments.