Add Smoke Tests
Adds a CodeBuild project that runs the forms-runner smoke test [1]. An
Event Bridge rule with a schedule is used to trigger the smoke tests.

CloudWatch metric alarms will trigger for when the smoke test is
failing or not running. The evaluation_period and threshold are to be
refined as we understand more about the requirements.

[1]: https://github.com/alphagov/forms-e2e-tests/pull/47