Consolidate .ruby-version files
We currently have multiple copies of the file `.ruby-version` in
multiple different places in this repo. This is mostly unnecssary, as
`rbenv` and similar tools will search up the folder heirarchy to find
the nearest file, and currently all these files specify the same version
of Ruby.

This commit adds a `.ruby-version` file to the root of the repo, and
removes two other copies of that file where it is expected that
the code will be run in an environment with the entire repository
checked out.

In cases where there is a directory heirarchy for a lambda function or
similar, we leave the `.ruby-version` file in place.

So there is some duplication still, but it is at least reduced.