Set Ruby version in GitHub Actions programmatically
We've moved the .ruby-version file to the root of the repository, but we still
have separate Gem files for separate code bases. The Ruby setup action assumes
that those two files will live in the same place, but that's not true now, so
we read the .ruby-version file and set the version from it explicitly.