Fix reference number feature flag variable name
The reference number feature flag variable was previously called
`SETTINGS__REFERENCE_NUMBERS_ENABLED`, but this doesn't work with our
settings gem. Because the value this flag overrides is
`Settings.features.reference_numbers_enabled`, the variable needs to be
`SETTINGS__FEATURES__REFERENCE_NUMBERS_ENABLED` for the gem to pick it
up correctly.