Use concurrent-ruby task to update pipeline data in the background
When we deployed the first iteration of the pipeline visualiser, it would hit API rate limits because it would make a flood of requests as a result of all the healthchecks pinging the "/" endpoint.

We could have implemented a less intensive healthcheck endpoint, but the
problem would have remained: under a reasonably small load, the thing
would hit API rate limits.