Rename Smoke Test Module to End-To-End
The existing browser tests in our pipeline are actually running the
end-to-end tests. We're about to introduce scheduled smoke tests so we
are renaming things to make the distinction clear.

This also specifies the Rspec file to run for the end-to-end tests so
that we do not inadvertently run the soon to be introduced smoke-tests
as well.