Allow human readonly roles in deploy to assume pipeline visualiser ECS task role
Running the pipeline visualiser requires being able to assume a set of roles.
This makes running it in development tricky, because we need to ensure that the
person running the code can assume all of the roles.

A convenient way of achieving that is to allow the humans to assume the role
that the code will run as when live. This ought to be OK because all of the
permissions involved as readonly and tightly scoped to the CodePipeline API.