Run a Rake task in `forms-admin` on a schedule
There is a Rake task in `forms-admin` which synchronises the users in the
database with mailing lists in MailChimp. We want to run this job daily to
ensure the two are kept in sync.

Scheduled tasks on AWS ECS use AWS EventBridge to trigger them using a
cron-like expression.

We only want this to be present in production, so we've added a new
configuration option to allow for that.