Allow dead letter queue config Checkov check to accept a string
We have a custom Checkov check that ensures that every AWS EventBridge target
has a dead letter queue configured. It previously did this by looking for a
connection between an aws_cloudwatch_event_target and an aws_sqs_queue. It now
allows for the value to be a string starting with "arn:aws:sqs", so that we can
hardcode the string when we want.