Add 'Error:' to Auth0 domain error text
When a user enters a non-government email address on the Auth0 login
page, they get an error message telling them to use their government
email address. The way Auth0 displays this message means that the only
way to tell it's an error message is a red background. This fails WCAG
1.4.1 Use of Color because there's nothing apart from colour to suggest
this is an error.

This commit prepends the error message with 'Error: ' so that even
without the colour it is obvious to users that this is an error that
they need to resolve.