Restore event notifications for the deploy account
When we refactored the event notifications, we began driving the accounts from
`local.other_accounts`. This list (correctly) doesn't contain the deploy
account, therefore we lost the notifications from that account.

This commit adds them back in by merging an extra key into that object. We
haven't added it to the `local.other_accounts` because that variable is used in
other places where having the deploy account would be wrong.