Add buildspec to perform db migration
This script is based on the db-migrations-pre-deploy.sh script [1]. It's job is
to run an ECS task that applies database migrations.
The main difference is that we do not have access to the task definitions and
network configuration in the same way, so we use api calls to grab those values.

In order to run database migrations before deploying a new version of an app, we
need:
- the existing task definition and configuration
- a new image uri
- an override for the comand that the container runs

We use `jq` to generate new task definition, and then run an ECS task based on
that definition.

[1] https://github.com/alphagov/forms-deploy/blob/f57d8c12d913161f3a361fd13b5beea0d6a0c61d/infra/modules/shared/scripts/db-migrations-pre-deploy.sh