Skip container image check for specific variable
We have checks to make sure we're not using multiple versions of the same image
in two places. They have picked up the variable for the image that gets passed
from codepipeline to the db-migrate buildspec, but we don't need it to check for
it here. To avoid the tests failing we have changed the tests to ignore this
specific variable