Change pipeline execution mode to queued
By default, pipelines are set to an execution mode of 'superseded', which means
that subsequent runs of the pipeline can supersede earlier runs if both are
queued at the same stage. We have changed this to 'queued', whereby all runs of
the pipeline are queued and run in order.

See here for more information:
https://docs.aws.amazon.com/codepipeline/latest/userguide/concepts-how-it-works.html#concepts-how-it-works-executions

Note that an execution mode of `queued` is only available for V2 pipelines.