Remove alerts module from alerts root
We want to move the content of the alerts module inside the environment
module, and remove the forms/alert root.

We could use Terraform's 'import' and 'removed' blocks to do migrate the
resoruces between the different roots' state files, but that's quite time
and labour intensive.

Instead, we're happy to suffer a few minutes outage among our alerts
whilst we:

1. Remove the resources from forms/alerts
2. Add the resources to the environment module.

This commit is step 1.